dashboard/requirements.txt
python-dotenv
pandas
pymssql
altair
streamlit